refactor and rename package
