+++ .github/workflows/bump.yaml
@@ -32,7 +32,7 @@ jobs:
         python-version: [ "3.9" ]
     runs-on: "${{ matrix.os }}"
     env:
-      CI_COMMIT_EMAIL: "ci-runner@ghg-forcing-for-cmip-comparison.invalid"
+      CI_COMMIT_EMAIL: "ci-runner@ghg-forcing-for-cmip.invalid"
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
