+++ README.md
@@ -1,27 +1,27 @@
 <!--- --8<-- [start:description] -->
-# GHG-forcing-for-CMIP-comparison
+# ghg-forcing-for-cmip
 
 Intercomparison of CMIP7 data with earth observations as part of GHG forcing for CMIP project.
 
 **Key info :**
-[![Docs](https://readthedocs.org/projects/ghg-forcing-for-cmip-comparison/badge/?version=latest)](https://ghg-forcing-for-cmip-comparison.readthedocs.io)
-[![Main branch: supported Python versions](https://img.shields.io/python/required-version-toml?tomlFilePath=https%3A%2F%2Fraw.githubusercontent.com%2Fclimate-resource%2FGHG-forcing-for-CMIP-comparison%2Fmain%2Fpyproject.toml)](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/blob/main/pyproject.toml)
-[![Licence](https://img.shields.io/pypi/l/ghg-forcing-for-cmip-comparison?label=licence)](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/blob/main/LICENCE)
+[![Docs](https://readthedocs.org/projects/ghg-forcing-for-cmip/badge/?version=latest)](https://ghg-forcing-for-cmip.readthedocs.io)
+[![Main branch: supported Python versions](https://img.shields.io/python/required-version-toml?tomlFilePath=https%3A%2F%2Fraw.githubusercontent.com%2Fclimate-resource%2Fghg-forcing-for-cmip%2Fmain%2Fpyproject.toml)](https://github.com/climate-resource/ghg-forcing-for-cmip/blob/main/pyproject.toml)
+[![Licence](https://img.shields.io/pypi/l/ghg-forcing-for-cmip?label=licence)](https://github.com/climate-resource/ghg-forcing-for-cmip/blob/main/LICENCE)
 
 **PyPI :**
 TODO
 <!--
-[![PyPI](https://img.shields.io/pypi/v/ghg-forcing-for-cmip-comparison.svg)](https://pypi.org/project/ghg-forcing-for-cmip-comparison/)
-[![PyPI install](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/actions/workflows/install-pypi.yaml/badge.svg?branch=main)](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/actions/workflows/install-pypi.yaml)
+[![PyPI](https://img.shields.io/pypi/v/ghg-forcing-for-cmip.svg)](https://pypi.org/project/ghg-forcing-for-cmip/)
+[![PyPI install](https://github.com/climate-resource/ghg-forcing-for-cmip/actions/workflows/install-pypi.yaml/badge.svg?branch=main)](https://github.com/climate-resource/ghg-forcing-for-cmip/actions/workflows/install-pypi.yaml)
 -->
 
 **Tests :**
-[![CI](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/actions/workflows/ci.yaml/badge.svg?branch=main)](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/actions/workflows/ci.yaml)
-[![Coverage](https://codecov.io/gh/climate-resource/GHG-forcing-for-CMIP-comparison/branch/main/graph/badge.svg)](https://codecov.io/gh/climate-resource/GHG-forcing-for-CMIP-comparison)
+[![CI](https://github.com/climate-resource/ghg-forcing-for-cmip/actions/workflows/ci.yaml/badge.svg?branch=main)](https://github.com/climate-resource/ghg-forcing-for-cmip/actions/workflows/ci.yaml)
+[![Coverage](https://codecov.io/gh/climate-resource/ghg-forcing-for-cmip/branch/main/graph/badge.svg)](https://codecov.io/gh/climate-resource/ghg-forcing-for-cmip)
 
 **Other info :**
-[![Last Commit](https://img.shields.io/github/last-commit/climate-resource/GHG-forcing-for-CMIP-comparison.svg)](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/commits/main)
-[![Contributors](https://img.shields.io/github/contributors/climate-resource/GHG-forcing-for-CMIP-comparison.svg)](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/graphs/contributors)
+[![Last Commit](https://img.shields.io/github/last-commit/climate-resource/ghg-forcing-for-cmip.svg)](https://github.com/climate-resource/ghg-forcing-for-cmip/commits/main)
+[![Contributors](https://img.shields.io/github/contributors/climate-resource/ghg-forcing-for-cmip.svg)](https://github.com/climate-resource/ghg-forcing-for-cmip/graphs/contributors)
 ## Status
 
 <!---
@@ -46,7 +46,7 @@ Some suggested options:
 <!--- --8<-- [end:description] -->
 
 Full documentation can be found at:
-[ghg-forcing-for-cmip-comparison.readthedocs.io](https://ghg-forcing-for-cmip-comparison.readthedocs.io/en/latest/).
+[ghg-forcing-for-cmip.readthedocs.io](https://ghg-forcing-for-cmip.readthedocs.io/en/latest/).
 We recommend reading the docs there because the internal documentation links
 don't render correctly on GitHub's viewer.
 
@@ -55,22 +55,22 @@ don't render correctly on GitHub's viewer.
 <!--- --8<-- [start:installation] -->
 ### As an application
 
-If you want to use GHG-forcing-for-CMIP-comparison as an application,
+If you want to use ghg-forcing-for-cmip as an application,
 then we recommend using the 'locked' version of the package.
 This version pins the version of all dependencies too,
 which reduces the chance of installation issues
 because of breaking updates to dependencies.
 
-The locked version of GHG-forcing-for-CMIP-comparison can be installed with
+The locked version of ghg-forcing-for-cmip can be installed with
 
 === "pip"
     ```sh
-    pip install 'ghg-forcing-for-cmip-comparison[locked]'
+    pip install 'ghg-forcing-for-cmip[locked]'
     ```
 
 ### As a library
 
-If you want to use GHG-forcing-for-CMIP-comparison as a library,
+If you want to use ghg-forcing-for-cmip as a library,
 for example you want to use it
 as a dependency in another package/application that you're building,
 then we recommend installing the package with the commands below.
@@ -78,18 +78,18 @@ This method provides the loosest pins possible of all dependencies.
 This gives you, the package/application developer,
 as much freedom as possible to set the versions of different packages.
 However, the tradeoff with this freedom is that you may install
-incompatible versions of GHG-forcing-for-CMIP-comparison's dependencies
+incompatible versions of ghg-forcing-for-cmip's dependencies
 (we cannot test all combinations of dependencies,
 particularly ones which haven't been released yet!).
 Hence, you may run into installation issues.
-If you believe these are because of a problem in GHG-forcing-for-CMIP-comparison,
-please [raise an issue](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/issues).
+If you believe these are because of a problem in ghg-forcing-for-cmip,
+please [raise an issue](https://github.com/climate-resource/ghg-forcing-for-cmip/issues).
 
-The (non-locked) version of GHG-forcing-for-CMIP-comparison can be installed with
+The (non-locked) version of ghg-forcing-for-cmip can be installed with
 
 === "pip"
     ```sh
-    pip install ghg-forcing-for-cmip-comparison
+    pip install ghg-forcing-for-cmip
     ```
 
 Additional dependencies can be installed using
@@ -97,10 +97,10 @@ Additional dependencies can be installed using
 === "pip"
     ```sh
     # To add plotting dependencies
-    pip install 'ghg-forcing-for-cmip-comparison[plots]'
+    pip install 'ghg-forcing-for-cmip[plots]'
 
     # To add all optional dependencies
-    pip install 'ghg-forcing-for-cmip-comparison[full]'
+    pip install 'ghg-forcing-for-cmip[full]'
     ```
 
 ### For developers
@@ -119,7 +119,7 @@ In order to create your environment, run `make virtual-environment`.
 
 If there are any issues, the messages from the `Makefile` should guide you through.
 If not, please raise an issue in the
-[issue tracker](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/issues).
+[issue tracker](https://github.com/climate-resource/ghg-forcing-for-cmip/issues).
 
 For the rest of our developer docs, please see [development][development].
