+++ docs/further-background/inspect_binned_data.py
@@ -14,10 +14,9 @@
 
 # %%
 import numpy as np
-import pandas as pd
 
-from ghg_forcing_for_cmip_comparison import plotting
-from ghg_forcing_for_cmip_comparison.utils import load_data
+from ghg_forcing_for_cmip.data_comparison import plotting
+from ghg_forcing_for_cmip.data_comparison.utils import load_data
 
 # %% [markdown]
 # # Inspect binned data
