+++ docs/installation.md
@@ -1,15 +1,15 @@
 # Installation
 
-There are a few different ways to install GHG-forcing-for-CMIP-comparison.
+There are a few different ways to install ghg-forcing-for-cmip.
 We provide these by use case below.
 As a short summary, if you:
 
-- just want to use GHG-forcing-for-CMIP-comparison for validation,
+- just want to use ghg-forcing-for-cmip for validation,
   and nothing else, go to [installation as an application][as-an-application]
-- want to use GHG-forcing-for-CMIP-comparison's functionality
+- want to use ghg-forcing-for-cmip's functionality
   as a library within another application,
   go to [installation as a library][as-a-library]
-- want to develop GHG-forcing-for-CMIP-comparison,
+- want to develop ghg-forcing-for-cmip,
   go to [installation for developers][for-developers]
 
 ## By use case
