+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "ghg-forcing-for-cmip-comparison"
+name = "ghg-forcing-for-cmip"
 version = "0.1.0a1"
 description = "Intercomparison of CMIP7 data with earth observations as part of GHG forcing for CMIP project."
 authors = [
@@ -41,11 +41,11 @@ classifiers = [
 ]
 
 [project.urls]
-Homepage = "https://ghg-forcing-for-cmip-comparison.readthedocs.io"
-Documentation = "https://ghg-forcing-for-cmip-comparison.readthedocs.io"
-Changelog = "https://ghg-forcing-for-cmip-comparison.readthedocs.io/en/stable/changelog"
-Repository = "https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison"
-Issues = "https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/issues"
+Homepage = "https://ghg-forcing-for-cmip.readthedocs.io"
+Documentation = "https://ghg-forcing-for-cmip.readthedocs.io"
+Changelog = "https://ghg-forcing-for-cmip.readthedocs.io/en/stable/changelog"
+Repository = "https://github.com/climate-resource/ghg-forcing-for-cmip"
+Issues = "https://github.com/climate-resource/ghg-forcing-for-cmip/issues"
 
 
 [project.optional-dependencies]
@@ -53,7 +53,7 @@ plots = [
     "matplotlib>=3.7.1",
 ]
 full = [
-    "ghg-forcing-for-cmip-comparison[plots]",
+    "ghg-forcing-for-cmip[plots]",
 ]
 
 [dependency-groups]
@@ -112,6 +112,7 @@ docs = [
     "mkdocstrings-python==1.13.0",
     "pymdown-extensions==10.13",
     "mkdocs-jupyter>=0.25.1",
+    "mkdocs-mermaid2-plugin>=1.2.1",
     "ruff==0.8.6",
     # Implied by the key dependencies above
     # -------------------------------------
@@ -278,7 +279,8 @@ build-backend = "pdm.backend"
 [tool.pdm.build]
 locked = true
 includes = [
-    "src/ghg_forcing_for_cmip_comparison",
+    "src/ghg_forcing_for_cmip_comparison/data_comparison",
+    "src/ghg_forcing_for_cmip_comparison/data_assimilation",
     "LICENCE",
 ]
 
@@ -383,13 +385,13 @@ package = "ghg_forcing_for_cmip_comparison"
 package_dir = "src"
 filename = "docs/changelog.md"
 directory = "changelog/"
-title_format = "## GHG-forcing-for-CMIP-comparison {version} ({project_date})"
+title_format = "## ghg-forcing-for-cmip {version} ({project_date})"
 underlines = [
     "",
     "",
     "",
 ]
-issue_format = "[#{issue}](https://github.com/climate-resource/GHG-forcing-for-CMIP-comparison/pull/{issue})"
+issue_format = "[#{issue}](https://github.com/climate-resource/ghg-forcing-for-cmip/pull/{issue})"
 type = [
     { directory = "breaking", name = "⚠️ Breaking Changes", showcontent = true },
     { directory = "deprecation", name = "🗑️ Deprecations", showcontent = true },
