+++ requirements-docs-locked.txt
@@ -44,13 +44,15 @@ babel==2.16.0
     #   jupyterlab-server
     #   mkdocs-material
 beautifulsoup4==4.12.3
-    # via nbconvert
+    # via
+    #   mkdocs-mermaid2-plugin
+    #   nbconvert
 bleach==6.2.0
     # via nbconvert
 cachetools==5.5.2
     # via prefect
 cdsapi==0.7.6
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 certifi==2024.12.14
     # via
     #   apprise
@@ -65,7 +67,7 @@ cffi==1.17.1
     #   cryptography
     #   pyzmq
 cftime==1.6.4.post1
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 charset-normalizer==3.4.1
     # via requests
 click==8.1.8
@@ -110,6 +112,8 @@ docker==7.1.0
     # via prefect
 ecmwf-datastores-client==0.2.0
     # via cdsapi
+editorconfig==0.17.1
+    # via jsbeautifier
 exceptiongroup==1.3.0
     # via
     #   anyio
@@ -128,11 +132,11 @@ fqdn==1.5.1
 fsspec==2025.5.1
     # via prefect
 geodatasets==2024.8.0
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 geopandas==1.0.1 ; python_full_version < '3.10'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 geopandas==1.1.1 ; python_full_version >= '3.10'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 ghp-import==2.1.0
     # via mkdocs
 graphviz==0.21
@@ -217,6 +221,8 @@ jinja2==3.1.5
     #   prefect
 jinja2-humanize-extension==0.4.0
     # via prefect
+jsbeautifier==1.15.4
+    # via mkdocs-mermaid2-plugin
 json5==0.10.0
     # via jupyterlab-server
 jsonpatch==1.33
@@ -295,11 +301,11 @@ markupsafe==3.0.2
     #   nbconvert
 matplotlib==3.9.4 ; python_full_version < '3.10'
     # via
-    #   ghg-forcing-for-cmip-comparison
+    #   ghg-forcing-for-cmip
     #   seaborn
 matplotlib==3.10.3 ; python_full_version >= '3.10'
     # via
-    #   ghg-forcing-for-cmip-comparison
+    #   ghg-forcing-for-cmip
     #   seaborn
 matplotlib-inline==0.1.7
     # via
@@ -322,6 +328,7 @@ mkdocs==1.6.1
     #   mkdocs-jupyter
     #   mkdocs-literate-nav
     #   mkdocs-material
+    #   mkdocs-mermaid2-plugin
     #   mkdocs-section-index
     #   mkdocstrings
 mkdocs-autorefs==1.2.0
@@ -337,6 +344,7 @@ mkdocs-material==9.5.49
     # via mkdocs-jupyter
 mkdocs-material-extensions==1.3.1
     # via mkdocs-material
+mkdocs-mermaid2-plugin==1.2.1
 mkdocs-section-index==0.3.9
 mkdocstrings==0.27.0
     # via mkdocstrings-python
@@ -431,13 +439,13 @@ paginate==0.5.7
 pandas==2.3.0
     # via
     #   geopandas
-    #   ghg-forcing-for-cmip-comparison
+    #   ghg-forcing-for-cmip
     #   seaborn
     #   xarray
 pandas-stubs==2.2.2.240807
     # via pandera
 pandera==0.24.0
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 pandocfilters==1.5.1
     # via nbconvert
 parso==0.8.4
@@ -453,7 +461,7 @@ pexpect==4.9.0 ; (python_full_version < '3.10' and sys_platform == 'emscripten')
 pillow==11.2.1
     # via matplotlib
 pip==24.3.1
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 platformdirs==4.3.6
     # via
     #   jupyter-core
@@ -463,7 +471,7 @@ platformdirs==4.3.6
 pooch==1.8.2
     # via geodatasets
 prefect==3.3.1
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 prometheus-client==0.21.1
     # via
     #   jupyter-server
@@ -506,6 +514,7 @@ pygments==2.19.1
 pymdown-extensions==10.13
     # via
     #   mkdocs-material
+    #   mkdocs-mermaid2-plugin
     #   mkdocstrings
 pyogrio==0.11.0
     # via geopandas
@@ -585,6 +594,7 @@ requests==2.32.3
     #   ecmwf-datastores-client
     #   jupyterlab-server
     #   mkdocs-material
+    #   mkdocs-mermaid2-plugin
     #   multiurl
     #   pooch
     #   requests-oauthlib
@@ -613,17 +623,19 @@ ruamel-yaml-clib==0.2.12 ; platform_python_implementation == 'CPython'
     # via ruamel-yaml
 ruff==0.8.6
 scipy==1.13.1 ; python_full_version < '3.10'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 scipy==1.15.3 ; python_full_version == '3.10.*'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 scipy==1.16.0 ; python_full_version >= '3.11'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 seaborn==0.13.2
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 send2trash==1.8.3
     # via jupyter-server
 setuptools==75.6.0
-    # via jupyterlab
+    # via
+    #   jupyterlab
+    #   mkdocs-mermaid2-plugin
 shapely==2.0.7 ; python_full_version < '3.10'
     # via geopandas
 shapely==2.1.1 ; python_full_version >= '3.10'
@@ -632,6 +644,7 @@ shellingham==1.5.4
     # via typer
 six==1.17.0
     # via
+    #   jsbeautifier
     #   python-dateutil
     #   rfc3339-validator
 sniffio==1.3.1
@@ -766,9 +779,9 @@ websockets==15.0.1
 whenever==0.7.3 ; python_full_version >= '3.13'
     # via prefect
 xarray==2024.7.0 ; python_full_version < '3.10'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 xarray==2025.6.1 ; python_full_version >= '3.10'
-    # via ghg-forcing-for-cmip-comparison
+    # via ghg-forcing-for-cmip
 zipp==3.23.0
     # via
     #   importlib-metadata
