+++ requirements-only-tests-min-locked.txt
@@ -1,10 +1,17 @@
 # This file was autogenerated by uv via the following command:
 #    uv export -o requirements-only-tests-min-locked.txt --no-hashes --no-dev --no-emit-project --only-group tests-min
 colorama==0.4.6 ; sys_platform == 'win32'
+    # via pytest
 exceptiongroup==1.3.0 ; python_full_version < '3.11'
+    # via pytest
 iniconfig==2.0.0
+    # via pytest
 packaging==24.2
+    # via pytest
 pluggy==1.5.0
+    # via pytest
 pytest==8.3.4
 tomli==2.2.1 ; python_full_version < '3.11'
+    # via pytest
 typing-extensions==4.12.2 ; python_full_version < '3.11'
+    # via exceptiongroup
