+++ src/ghg_forcing_for_cmip/__init__.py
@@ -1,9 +1,9 @@
 """
-Data intercomparison
+GHG forcing for CMIP
 
 Compare CMIP7 data with earth observations as part of GHG forcing for CMIP project.
 """
 
 import importlib.metadata
 
-__version__ = importlib.metadata.version("ghg_forcing_for_cmip_comparison")
+__version__ = importlib.metadata.version("ghg_forcing_for_cmip")
