+++ src/ghg_forcing_for_cmip/data_comparison/CONFIG.py
