+++ src/ghg_forcing_for_cmip/data_comparison/__init__.py
@@ -1,0 +1,5 @@
+"""
+Data intercomparison
+
+Compare CMIP7 data with earth observations as part of GHG forcing for CMIP project.
+"""
