+++ src/ghg_forcing_for_cmip/data_comparison/combine_datasets.py
@@ -5,8 +5,8 @@ combine earth observations and ground-based data in one dataset
 import pandas as pd
 from prefect import flow, task
 
-from ghg_forcing_for_cmip_comparison import CONFIG
-from ghg_forcing_for_cmip_comparison.utils import save_data
+from ghg_forcing_for_cmip.data_comparison import CONFIG
+from ghg_forcing_for_cmip.data_comparison.utils import save_data
 
 
 @task(
