+++ src/ghg_forcing_for_cmip/data_comparison/get_datasets.py
@@ -16,7 +16,7 @@ import requests
 import xarray as xr
 from prefect import flow, task
 
-from ghg_forcing_for_cmip_comparison import CONFIG, utils
+from ghg_forcing_for_cmip.data_comparison import CONFIG, utils
 
 
 @task(
