+++ src/ghg_forcing_for_cmip/data_comparison/interpolate_dataset_gb.py
@@ -14,7 +14,7 @@ import xarray as xr
 from prefect import flow, task
 from scipy.interpolate import griddata  # type: ignore
 
-from ghg_forcing_for_cmip_comparison import CONFIG
+from ghg_forcing_for_cmip.data_comparison import CONFIG
 
 
 @task(
