+++ src/ghg_forcing_for_cmip/data_comparison/workflow.py
@@ -4,10 +4,14 @@ Main workflow
 
 from prefect import flow
 
-from ghg_forcing_for_cmip_comparison.bin_dataset_gb import bin_dataset_flow
-from ghg_forcing_for_cmip_comparison.get_datasets import get_data_flow
-from ghg_forcing_for_cmip_comparison.interpolate_dataset_gb import interpolation_flow
-from ghg_forcing_for_cmip_comparison.vertical_to_dataset_gb import add_vertical_flow
+from ghg_forcing_for_cmip.data_comparison.bin_dataset_gb import bin_dataset_flow
+from ghg_forcing_for_cmip.data_comparison.get_datasets import get_data_flow
+from ghg_forcing_for_cmip.data_comparison.interpolate_dataset_gb import (
+    interpolation_flow,
+)
+from ghg_forcing_for_cmip.data_comparison.vertical_to_dataset_gb import (
+    add_vertical_flow,
+)
 
 
 @flow(
@@ -18,7 +22,7 @@ def run_pipeline_comparison(
     gas: str, quantile: float, save_to_path: str = "data/downloads"
 ) -> None:
     """
-    Run main workflow for GHG-forcing-for-CMIP-comparison
+    Run main workflow for ghg-forcing-for-cmip
     """
     get_data_flow(save_to_path=save_to_path)
