+++ tests/unit/test_download_datasets.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from ghg_forcing_for_cmip_comparison.get_datasets import (
+from ghg_forcing_for_cmip.data_comparison.get_datasets import (
     download_noaa,
     unzip_download,
 )
